+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
 from framewirc import exceptions
-from framewirc.utils import chunk_message, RequiredAttributesMixin, to_bytes, to_unicode
+from framewirc.utils import RequiredAttributesMixin, to_bytes, to_unicode
 
 
 class TestToUnicode:
@@ -61,58 +61,6 @@ class TestToBytes:
     def test_not_bytes_or_string(self):
         with pytest.raises(TypeError):
             to_bytes(None)
-
-
-class TestChunkMessage:
-    """Test the behaviour of the chunk_message function."""
-    def test_return_type(self):
-        """Does it return a list of bytes objects?"""
-        messages = chunk_message('Just a simple message', max_length=100)
-        assert messages == [b'Just a simple message']
-
-    def test_split_linefeeds(self):
-        """Does it split on newline chars?"""
-        msg = 'A message\rsplit over\nmany lines\r\nwith odd linebreaks.'
-        expected = [
-            b'A message',
-            b'split over',
-            b'many lines',
-            b'with odd linebreaks.',
-        ]
-        assert chunk_message(msg, max_length=100) == expected
-
-    def test_split_long_line(self):
-        """Does it split long lines?"""
-        msg = 'Message to be split into chunks of twenty characters or less.'
-        expected = [
-            b'Message to be split',
-            b'into chunks of',
-            b'twenty characters or',
-            b'less.',
-        ]
-        assert chunk_message(msg, max_length=20) == expected
-
-    def test_split_long_word(self):
-        """Does it split long words?"""
-        msg = 'Sup Llanfairpwllgwyngyllgogerychwyrndrobwllllantysiliogogogoch?'
-        expected = [
-            b'Sup',
-            b'Llanfairpwllgwyngyll',
-            b'gogerychwyrndrobwlll',
-            b'lantysiliogogogoch?',
-        ]
-        assert chunk_message(msg, max_length=20) == expected
-
-    def test_split_long_unicode(self):
-        """Are words with multi-byte chars split correctly?"""
-        # Repeated failures lead to success.
-        msg = '失敗を繰り返すことで、成功に至る。'
-        expected = [
-            to_bytes('失敗を繰り返'),
-            to_bytes('すことで、成'),
-            to_bytes('功に至る。'),
-        ]
-        assert chunk_message(msg, max_length=20) == expected
 
 
 class TestRequiredAttributesMixin: